+++ grpc/src/main/java/com/linecorp/armeria/client/grpc/endpoint/healthcheck/GrpcHealthCheckWatcher.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package com.linecorp.armeria.internal.client.grpc;
+package com.linecorp.armeria.client.grpc.endpoint.healthcheck;
 
 import static java.util.Objects.requireNonNull;
 
@@ -41,7 +41,7 @@ import io.grpc.stub.StreamObserver;
 /**
  * Performs gRPC health checking using the Watch rpc endpoint.
  */
-public class GrpcHealthCheckWatcher extends AbstractGrpcHealthChecker {
+class GrpcHealthCheckWatcher extends AbstractGrpcHealthChecker {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(GrpcHealthCheckWatcher.class);
 
@@ -51,8 +51,8 @@ public class GrpcHealthCheckWatcher extends AbstractGrpcHealthChecker {
     @Nullable
     private ClientRequestContext activeRequestContext;
 
-    public GrpcHealthCheckWatcher(HealthCheckerContext ctx, Endpoint endpoint, SessionProtocol sessionProtocol,
-                                  @Nullable String service) {
+    GrpcHealthCheckWatcher(HealthCheckerContext ctx, Endpoint endpoint, SessionProtocol sessionProtocol,
+                           @Nullable String service) {
         this.ctx = requireNonNull(ctx, "ctx");
         requireNonNull(endpoint, "endpoint");
         requireNonNull(sessionProtocol, "sessionProtocol");
@@ -78,7 +78,8 @@ public class GrpcHealthCheckWatcher extends AbstractGrpcHealthChecker {
             }
 
             try (ClientRequestContextCaptor reqCtxCaptor = Clients.newContextCaptor()) {
-                StreamObserver<HealthCheckResponse> responseObserver = new StreamObserver<HealthCheckResponse>() {
+                final StreamObserver<HealthCheckResponse> responseObserver =
+                        new StreamObserver<HealthCheckResponse>() {
                     @Override
                     public void onNext(HealthCheckResponse healthCheckResponse) {
                         lock();
